Optimizations

See merge request ItayS14/carmel-1703-voice-assistant!17

+++ Server/models.py
@@ -17,7 +17,7 @@ class User(db.Model, UserMixin):
     password = db.Column(db.String(60), nullable=False)
     house_id = db.Column(db.Integer, db.ForeignKey('house.id'))
     reset_code = db.Column(db.String(6), unique=True, nullable=True)
-    updated_time = db.Column(db.Integer, nullable=True)
+    updated_time = db.Column(db.Float, nullable=True)
     profile_image = db.Column(db.String(32), nullable=False, default='default.jpg')
     confirmed = db.Column(db.Boolean, nullable=False, default=False)
